+++ src/Kafka/KafkaInstanceDrawer/KafkaSampleCode.stories.tsx
@@ -4,9 +4,7 @@ import { KafkaSampleCode } from "./KafkaSampleCode";
 
 export default {
   component: KafkaSampleCode,
-  args: {
-    value: "java",
-  },
+  args: {},
   parameters: {
     backgrounds: {
       default: "Background color 100",
@@ -18,19 +16,4 @@ const Template: ComponentStory<typeof KafkaSampleCode> = (args) => (
   <KafkaSampleCode {...args} />
 );
 
-export const JavaSampleCode = Template.bind({});
-
-export const PythonSampleCode = Template.bind({});
-PythonSampleCode.args = {
-  value: "python",
-};
-
-export const QuarkusSampleCode = Template.bind({});
-QuarkusSampleCode.args = {
-  value: "quarkus",
-};
-
-export const SpringbootSampleCode = Template.bind({});
-SpringbootSampleCode.args = {
-  value: "springboot",
-};
+export const SampleCodeSnippet = Template.bind({});
