+++ src/Kafka/KafkaInstanceDrawer/components/useCopyClipBoard.ts
@@ -1,0 +1,21 @@
+import { useState } from "react";
+
+export function useCopyClipBoard() {
+  const [copied, setCopied] = useState<boolean>(false);
+
+  const clipboardCopyFunc = (event: any, text: string) => {
+    navigator.clipboard
+      .writeText(text.toString())
+      .then(() => {
+        setCopied(true);
+      })
+      .catch(() => {
+        setCopied(false);
+      });
+  };
+
+  return {
+    copied,
+    clipboardCopyFunc,
+  };
+}
