+++ src/Kafka/KafkaInstanceDrawer/types.ts
@@ -1,5 +1,1 @@
 export type ClientType = "java" | "python" | "quarkus" | "springboot";
-
-export type CodeSnippetType = "config" | "consumer" | "producer";
-
-export type SpringBootConsumerType = "consumerConfig" | "listener" | "consumer";
